+++ examples/gh-pages/scripts/client.js
@@ -1,11 +1,10 @@
 import {default as React} from "react";
+import {default as ReactDOM} from "react-dom";
 
-import {default as ReactRoot} from "./ReactRoot";
-
-require("../styles/index.scss");
+import {default as SyncReactRoot} from "./containers/SyncReactRoot";
 
 /*
  *
  * Add <script src="https://maps.googleapis.com/maps/api/js"></script> to your HTML to provide google.maps reference
  */
-React.render(<ReactRoot />, document.getElementById("react-container"));
+ReactDOM.render(<SyncReactRoot />, document.getElementById("react-container"));