+++ examples/gh-pages/scripts/components/GettingStarted.js
@@ -1,9 +1,8 @@
-import {default as React, addons, Component} from "react/addons";
+import {default as React, Component} from "react";
+import {default as update} from "react-addons-update";
 
 import {default as GoogleMap} from "../../../../src/GoogleMap";
 import {default as Marker} from "../../../../src/Marker";
-
-const {update} = addons;
 
 /*
  * This is the modify version of:
@@ -28,7 +27,7 @@ export default class GettingStarted extends Component {
    * This is called when you click on the map.
    * Go and try click now.
    */
-  _handle_map_click = (event) => {
+  _handle_map_click (event) {
     var {markers} = this.state;
     markers = update(markers, {
       $push: [
@@ -75,7 +74,7 @@ export default class GettingStarted extends Component {
         ref="map"
         defaultZoom={3}
         defaultCenter={{lat: -25.363882, lng: 131.044922}}
-        onClick={this._handle_map_click}>
+        onClick={::this._handle_map_click}>
         {this.state.markers.map((marker, index) => {
           return (
             <Marker