+++ examples/gh-pages/scripts/containers/NavHeaderBar.js
@@ -1,4 +1,9 @@
-import {default as React, PropTypes, Component} from "react";
+import {
+  default as React,
+  Component,
+  PropTypes,
+} from "react";
+
 import {default as cx} from "classnames";
 
 const actionPropType = PropTypes.shape({
@@ -67,7 +72,7 @@ export default class NavHeaderBar extends Component {
               <li><a href="https://github.com/tomchentw" target="_blank">by @tomchentw</a></li>
               {props.actions.map(actionToMenuItem, this)}
               <li className={cx(dropdownClassSet)}>
-                <a href="javascript:void(0);" className="dropdown-toggle" onClick={this._handle_click.bind(this)}>Samples <span className="caret"></span></a>
+                <a href="javascript:void(0);" className="dropdown-toggle" onClick={::this._handle_click}>Samples <span className="caret"></span></a>
                 <ul className="dropdown-menu" role="menu">
                   {props.dropdownActions.map(actionToMenuItem, this)}
                 </ul>