+++ src/addons/InfoBox.js
@@ -4,6 +4,10 @@ import {
 } from "react";
 
 import {
+  default as canUseDOM,
+} from "can-use-dom";
+
+import {
   default as InfoBoxCreator,
   infoBoxDefaultPropTypes,
   infoBoxControlledPropTypes,
@@ -41,9 +45,11 @@ export default class InfoBox extends Component {
   state = {
   }
 
-  componentDidMount () {
-    const {mapHolderRef, ...infoBoxProps} = this.props;
-    const infoBox = InfoBoxCreator._createInfoBox(mapHolderRef, infoBoxProps);
+  componentWillMount () {
+    if (!canUseDOM) {
+      return;
+    }
+    const infoBox = InfoBoxCreator._createInfoBox(this.props);
 
     this.setState({ infoBox });
   }