+++ src/addons/addonsCreators/InfoBoxCreator.js
@@ -14,6 +14,10 @@ import {default as componentLifecycleDecorator} from "../../utils/componentLifec
 import {default as GoogleMapHolder} from "../../creators/GoogleMapHolder";
 
 export const infoBoxControlledPropTypes = {
+// NOTICE!!!!!!
+//
+// Only expose those with getters & setters in the table as controlled props.
+//
 // http://google-maps-utility-library-v3.googlecode.com/svn/trunk/infobox/docs/reference.html
   content: PropTypes.any,
   options: PropTypes.object,
@@ -25,7 +29,7 @@ export const infoBoxControlledPropTypes = {
 export const infoBoxDefaultPropTypes = defaultPropsCreator(infoBoxControlledPropTypes);
 
 const infoBoxUpdaters = {
-  children  (children, component) { setContentForOptionalReactElement(children, component.getInfoWindow()); },
+  children  (children, component) { setContentForOptionalReactElement(children, component.getInfoBox()); },
   content   (content, component) { component.getInfoBox().setContent(content); },
   options   (options, component) { component.getInfoBox().setOptions(options); },
   position  (position, component) { component.getInfoBox().setPosition(position); },
@@ -49,20 +53,15 @@ export default class InfoBoxCreator extends Component {
     infoBox: PropTypes.object.isRequired,
   }
 
-  static _createInfoBox (mapHolderRef, infoBoxProps, anchorHolderRef) {
+  static _createInfoBox (infoBoxProps) {
+    const {mapHolderRef, anchorHolderRef} = infoBoxProps;
     // "google-maps-infobox" uses "google" as a global variable. Since we don't
     // have "google" on the server, we can not use it in server-side rendering.
     // As a result, we import "google-maps-infobox" here to prevent an error on
     // a isomorphic server.
     const GoogleMapsInfobox = require("google-maps-infobox");
     // http://google-maps-utility-library-v3.googlecode.com/svn/trunk/infobox/docs/reference.html
-    const infoBox = new GoogleMapsInfobox(composeOptions(infoBoxProps, [
-      // https://developers.google.com/maps/documentation/javascript/3.exp/reference
-      "content",
-      "position",
-      "visible",
-      "zIndex",
-    ]));
+    const infoBox = new GoogleMapsInfobox(composeOptions(infoBoxProps, infoBoxControlledPropTypes));
 
     if (infoBoxProps.children) {
       setContentForOptionalReactElement(infoBoxProps.children, infoBox);