+++ src/creators/MarkerCreator.js
@@ -14,7 +14,12 @@ import {default as componentLifecycleDecorator} from "../utils/componentLifecycl
 import {default as GoogleMapHolder} from "./GoogleMapHolder";
 
 export const markerControlledPropTypes = {
+// NOTICE!!!!!!
+//
+// Only expose those with getters & setters in the table as controlled props.
+//
 // [].map.call($0.querySelectorAll("tr>td>code", function(it){ return it.textContent; }).filter(function(it){ return it.match(/^set/) && !it.match(/^setMap/); })
+//
 // https://developers.google.com/maps/documentation/javascript/3.exp/reference#Marker
   animation: PropTypes.any,
   attribution: PropTypes.any,
@@ -69,25 +74,10 @@ export default class MarkerCreator extends Component {
     marker: PropTypes.object.isRequired,
   }
 
-  static _createMarker (mapHolderRef, markerProps) {
+  static _createMarker (markerProps) {
+    const {mapHolderRef} = markerProps;
     // https://developers.google.com/maps/documentation/javascript/3.exp/reference#Marker
-    const marker = new google.maps.Marker(composeOptions(markerProps, [
-      // https://developers.google.com/maps/documentation/javascript/3.exp/reference#MarkerOptions
-      "animation",
-      "attribution",
-      "clickable",
-      "cursor",
-      "draggable",
-      "icon",
-      "label",
-      "opacity",
-      "place",
-      "position",
-      "shape",
-      "title",
-      "visible",
-      "zIndex",
-    ]));
+    const marker = new google.maps.Marker(composeOptions(markerProps, markerControlledPropTypes));
 
     marker.setMap(mapHolderRef.getMap());
 
@@ -110,7 +100,7 @@ export default class MarkerCreator extends Component {
     if (0 < Children.count(children)) {
       return (
         <div>{Children.map(children, childElement =>
-          React.cloneElement(childElement, {
+          childElement && React.cloneElement(childElement, {
             mapHolderRef,
             anchorHolderRef: this,
           })